Allow change max temperature
Currently only way to change max temperature is use dbus message.
But the platform where there is no dbus, then also this change
allows setting on max temperarture via xml config file.

For example if for "CPU" zone if the calculated max temperature
is 90000, but it is desired only 80000, then the following
config will change this:

<ThermalConfiguration>
<Platform>
	<Name>Generic X86 Laptop Device</Name>
	<ProductName>*</ProductName>
	<Preference>QUIET</Preference>
	<ThermalZones>
		<ThermalZone>
			<Type>cpu</Type>
			<TripPoints>
				<TripPoint>
					<Temperature>80000</Temperature>
					<type>max</type>
				</TripPoint>
			</TripPoints>
		</ThermalZone>
	</ThermalZones>
</Platform>
</ThermalConfiguration>