When polling temp offset when max is present
This address the issue when the zone has max temperature and a
passive temp which is lower than polling threshold calculated from
max temp. Then polling temp becomes passive temp value. When we are
below polling temp, we withdraw cooling compensation, this will
cause zig-zag effect. So we add offset in all cases when passive
temp is deciding polling point.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>