By default don't activate surface zone
There is some over compensation reports because of wrong skin
temperarures. The problem is that the sensor is not mounted on
skin so actual temperature on lsptop skin may be much lower.
This control can always be activated via thermal-conf.xml.
So added default config in thermal-conf.xml, so that users
can change without touching code.