Null terminate buffer on error return path
Coverity Scan is being rather too paranoid about the error return
path not null terminating a string. This is a false positive in the
code, but it does make some sense to ensure the string is null
terminated in case future callers to csys_fs::read_symbolic_link_value
are sloppy in their error return checking.

Signed-off-by: Colin Ian King <colin.king@canonical.com>