Fix indexes
When thermal zones are contiguous (has holes) then the current
index for a zone/cdev/sensor has a bug. It is depending on the
count for next index. But when there is hole, this count will not
be next index and will cause duplicates.
Also removed using variable for count, instead use .size() to
avoid bad count.