Fix null pointer deferences with _ART active trip points (LP: #1480821)
We've been seeing null pointer deferences on some Intel devices that
support _TRT and _ART ACPI objects.  This is due to a typo in
parse_target_devices() where a non-existant trt object is being pushed
onto a trt vector instead of the art vector.  This causes a NULL pointer
deference in add_active_trip_point() when performing the following:

  union art_object *object = (union art_object *) rel_obj.art_objects[j];

Signed-off-by: Colin Ian King <colin.king@canonical.com>