Re-enable lockfile (LP:# 1480892)
The lock file is no longer working, this was caused by adding
a close() on the lock file fd on a CoverityScan cleanup on
commit 97b1720a42f33fd21e6fbf30b6b3857a948e5633
("Fix Coverity error in release").  This fix re-works the lockfile
so that it is only closed on thermal exit so that the lock is held
during the lifetime of thermald, and also the lock file is unlinked
on thermald exit to clean itself up properly.

Signed-off-by: Colin Ian King <colin.king@canonical.com>