csys_fs::write: return -errno on write failure
sysfs writes can fail for various reasons, and returning -errno is
useful for indicating an error so that the caller can determine
the best strategy to take if a sysfs write fails.  This change
is compatible with the current semantics, e.g. -ve is an error, but
also propagates the errno back to the caller.

Signed-off-by: Colin Ian King <colin.king@canonical.com>