thermald: fix incorrect target state settting by thermald
It was observed that in the case of cooling device with min_state > max_state,
thermald doesn't set the cooling device to the intended target state per the
configuration in XML. This is because of sorting the zone_trip_limits list in
ascending order based on target value is not suitable for the case of cooling
device with min_state > max_state. Thus this issue is fixed by:

1. sorting the zone_trip_limits list in descending order based on target value
for the case of cooling device with min_state > max_state.

2. sorting the zone_trip_limits list in ascending order based on target value
for the case of cooling device with max_state > min_state.

Example cooling device configuration which caused this issue:
<CoolingDevice>
	<Type>gpu_cooling</Type>
	<Path>/sys/class/mali/pm/max_freq_level</Path>
	<MinState>4</MinState>
	<MaxState>1</MaxState>
</CoolingDevice>.

Signed-off-by: Prashanth Voora <prashanth.voora@intel.com>