Fix multiple trip active for a cooling device
It is possible that same zone or different zone has two trips active
for a cooling device. There was a bug when trip id matches from
two different zones and one gets deactivated. For example
-Zone 1: trip 0 activated a cooling device
-Zone 2: trip 0 activated same cooling device
With the above any one trip deativated the trip mask will be reset.
With this change zone id and trip id combined added to the list
which was used only when a target value is set. When a trip id is
deactivated it will only remove its as zone id and trip id comination
will be always unique.