Fix for bad phy_max settings
When the FW has the bad constraint_0_max_power_uw, it may result in
negative max power limit when casted to integer max state.
Changed all the power limits to int from long as the integer can hold
up to 2147 Watts.
With this any bad value of phy_max which results in value < 0 or
more than 100W, we will simply ignore the max and calculate dynamically.