New tag WritePrefix for cooling device
Currently user can write only a integer value in a sysfs path
for a cooling device. But sometimes some string needs to be
prefixed. For example:
If user has to do
echo level 1 > some_path
This can't be defined in the xml config.

With the new tag WritePrefix, "level " will be always
prefixed in front of the state.

<CoolingDevice>
    <Type>_fan_</Type>
    <Path>/proc/acpi/ibm/fan</Path>
    <WritePrefix>level </WritePrefix>
    <MinState>0</MinState>
    <MaxState>5</MaxState>
    <DebouncePeriod>10</DebouncePeriod>
</CoolingDevice>