Optimize mutistep, multizone control
Suppose these are the steps in a zone.

Zone 0 : Sensor: sensor_1
	temp: 35 target_state: 15W
	temp: 40 target_state: 10W
	temp: 45 target_state: 8W

Zone 1 : Sensor: sensor_2
	temp: 35 target_state: 15W
	temp: 40 target_state: 10W
	temp: 45 target_state: 8W

In the current code:

Sequence:
sensor_1 == 45
   then zone 0 will cause target state to 8

sensor_2 == 40
   Since current power 8W is more constrained than 10W
   this request will not cause call to thd_cdev_set_state()

sensor_1 == 35
   then zone 0 will cause target state to 15W

But since sensor_2 is still == 40, but the power will be
10W, more than 8W for couple of seconds till another
scan cycle.

This change will still call thd_cdev_set_state for zone_1:
sensor_2, the request will be queued in zone_trip_limits().
But will not be activates as it is higher than the current
limit. But when sensor_1 temp == 35, the next entry in the
list will be picked up which will be 10W here without couple
of seconds of 15W.