Add tag for dmi product_sku
Add a "Product SKU" tag, so that this can be matched with dmi information
product_sku, and uniquely identify a platform in addition to "ProductName"
For example:
<Platform>
        <Name> Auto generated </Name>
        <ProductName>XPS 13 7390 2-in-1</ProductName>
        <ProductSku>08B0</ProductSku>
        <Preference>QUIET</Preference>
        <PPCC>
...
...

Here ProductName and ProductSku is matched with DMI information
to identify configuration for a platform. The ProductSku is optional
and also can be a wildcard "*".