Special handling to remove full control
[INFO]Zone 9: auto_zone_1, Active:1 Bind:1 Sensor_cnt:1
[INFO]..sensors..
[INFO]sensor index:1 B0D4 /sys/class/thermal/thermal_zone1/ Async:0
[INFO]..trips..
[INFO]index 0: type:passive temp:51000 hyst:0 zone id:9 sensor id:1 control_type:0 cdev size:1
[INFO]cdev[0] rapl_controller, Sampling period: 0
[INFO]	 target_state:2147483647
[INFO]index 1: type:passive temp:52000 hyst:0 zone id:9 sensor id:1 control_type:0 cdev size:1
[INFO]cdev[0] rapl_controller, Sampling period: 5
[INFO]	 target_state:not defined
[INFO]index 2: type:polling temp:45900 hyst:0 zone id:9 sensor id:1 control_type:0 cdev size:0

Here it is expected that at 52C, full control is applied but at 51C full power is restored.
But here there is an issue.

if temp > 52 then
  there will be two entries in the list
	entry 0: temperature 51
	entry 1: temperature 52

Then temp is changed < 51

First thd_cdev_set_state will be called for 51C with state = 0
	this will cause
		entry 0 will be deleted

This will result in control will be restored using exponential controller stepwise, as there
is no entry in the queue for 51C in the list.

The fix is here to delete the entry which was currently used. In this way first the control
for 52C will be removed and then the next entry is 51C, so full power is restored at 51C.