Ignore invalid PPCC max power limit
Some newer platforms moved the PPCC power limits to a new DPTF
configuration file. But they left junk at the
/sys/bus/pci/devices/0000\:00\:04.0/power_limits/power_limit_0_max_uw

In Dell 7390 the default max is 15W, but the value at the above
sysfs entry is 9W, causing performance loss. In this case simply
ignore all the data from the above sysfs entry.

But if dptfxtract is ran before then this sysfs value will be ignored
as we have this limit from DPTF config files. As a safety, added a
warning if that value is also less than the default.