Implement fallback mode with --adaptive option
Unless a distro ships with --adaptive option it is not likely that many users
will use this. But since this option may fail, implement some fallback.

Fallback method in order:

- It is likely that unsupported conditions will exist. So in this
case it is possible that some existing conditions are met, so
evaluate and if successful, then continue to use this method.

- If the above condition fails, then pick a target with highest PL1MAX.

- If the above condition fails then fallback to old method of using
XML, _TRT or default in built trips.