Delete all trip and add again
With exclusive option the zones and trips are created from thermal
sysfs. This may have junk in systems with adative policies. So
simply delete all existing trips created from the sysfs. Recreate
as part of install_passive(). This will also add multiple trips.

Also when the target table is changed, this will recreate with
new new entries.

It is possible that while in the middle of throttling where
a cooling device is not in min state and adpative mode change. In
this case zone_reset() is called but since the old min_state is
not the new one, it will not reset the cooling device. So use
force argument for zone_reset().