Fix abort on Zbook HP ZBook 15
terminate called after throwing an instance of 'std::invalid_argument'
  what():  stoi
Aborted

The problem is that conditions comparison for names failed for
case comparison, so it was passing string to to convert to stoi().
This generates exception.

To avoid this use try and catch and also use ignore case during
comparison.