cdev: Initialize PID at unique trip-point to current value
The PID regulator needs to start with a reasonable value. Just starting
at the minimum, would result in an immediate offset. The problem with
the current code was that it would always add the current state,
resulting in an integrative behaviour which will result in an overshoot
towards setting the maximum state for the cdev after a short period of
time.

Closes: #272