Deal case where target.participant is empty
Instead of totally fail to install PSVT and use limits from  sysfs
use target.code == "PSVT" to install. In one Lenovo system, this
is the case. Also sysf CPU temperature threshold is too low, which
is not tested by OEM, so throttled to much lower value. So backup
threshold also can't be used.