Fix error for condition names
The current code caps the max name as the last condition name,
which is "Power_Slider". So any condition more than 56 will be
printing error, with "Power_Slider" as condition name. For example
for condition = 57:
Unsupported condition 57 (Power_slider)

This is confusing during debug, so print "UNKNOWN" for condition
name 56.