AC/DC plug unplug issue in HP TGL laptop
It is possible that the EC lowers the limit on charger unplug. Suppose
the AC PL1MAX is 9W and on unplug charger the PL1 is changed by EC to 7W,
the current logic will not update the PL1 if the adaptive table calls for
9W for DC also. Since the current state when unthrottled will be still 9W
on unplug and the new state calls for 9W also, if the EC didn't change the
 limit underneath, it will fine.

But here EC is also changing to PL1 limit underneath, the system will
be running on reduced performance. So update the current state to
reflect new PL1 value set by EC.

If the new state is more than the desired value from the PL1MAX from
adaptive table, it will be used as min state.