adaptive: Implement Power Slider following power-profiles-deamon
This maps the three power-profiles-daemon states to appropriate values
of the power slider DPTF control.

Note that thermald already has the "preference" setting which can be set
to either "PERFORMANCE" or "ENERGY_CONSERVE". As such, an alternative
could be to expand the list of valid preferences and then let p-p-d set
the value via dbus.