engine: Split start into init/start to load sensors earlier
The adaptive engine needs to have the sensors loaded early on in order
to verify conditions. To make sure this works well, split the start
function into two parts, the first being init and the second being
start.

For adaptive, the start function still does some verification, as the
upower availability is important for the lid state (it should switch to
reading the lid state directly though).