Install ITMT target
Similar to PSVT install, implement ITMT install. When APAT calls
for ITMT instead of ITMT install ITMT target. The install portion
is much simpler. There can be only one trip in a ITMT entry so
no need to consolidate.

An ITMT entry will have a trip, min and max value for RAPL PL1.
So create trip with this information and bind RAPL cooling device
to this.