main: Handle signals from mainloop
The signal handler ends up calling quite a lot of code, not all of this
code is signal safe. Fix the issue by instead using the GLib
infrastructure to execute the code from the mainloop instead.

Note that there are probably more cleanups to be had here. But this is
sufficient to fix the signal safety.

Fixes: #361