Manual: thermal-conf.xml - clean up lintian warnings and .IP formatting
The debian packaging lintian checker is complaining that some of the
lines in the manual are very long. Break these up to sub-80 char wide
lines to clean these warnings up.  Also, fix up the .IP formatting as
the intentation is a bit broken. Add bullet points in .IP formatting
for some style too.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>