Workaround for missing PSVT or ITMT target
When there is no PSVT or ITMT target is specified, currently it
fallbacks to the sysfs defaults for CPU thermal zone. It is possible
that those CPU sysfs entries are never validated by the OEM.

So fallback to IETM_D0 table if present. Since this table is in the
GDDV, hopefully this is correct.