Install passive when target fails to match
It is possible that some OEM variable conditions are present, this
is a special case as this will always match during init. So there
will not be any fallback even if later none of the conditions
with OEM variable match.

In this case during periodic call to update_engine_state() install
IETM D0 default entry.