Fix seg fault when there is no config file
thermald currently segfaults on its first run.

If a configuration file is missing, `cthd_acpi_rel::generate_conf`
will invoke `cthd_acpi_rel::parse_target_devices` without a prior call
to `cthd_acpi_rel::process_psvt`, which is the only place that
initializes the `psvt_data` and `psvt_count` variables and both are
read uninitialized, which results in a segfault. Unlike its art and
trt counterparts, the two variables are not initialized in the
constructor.