Coverity issue for oveflowed constant
Silence coverity error for overflowed constant:
overflow_const: Expression this->zone_trip_limits.size() - 1UL,
which is equal to 18446744073709551615, where
this->zone_trip_limits.size() is known to be equal to 0,
underflows the type that receives it, an unsigned integer 64 bits wide.

This situation can't happen, but reduce Coverity noise.