Fix issue with fast polling and polling disabled
Even if polling mode is disabled, polling is still used to get
temperature when it is close to the trip. This is done because
the async notifications may be delayed for change.

There is a logical issue in this flow.

When first time the polling trip is reached, the poll interval
is default poll interval (4 sec). But once trip is breached then
polling switches to fast polling of one second.

When switched to fast polling, the last poll interval is stored
to restore later once fast polling is turned OFF. But if
the temperature dropped below the poll trip, the slow polling
disable can be called first, which changes to poll interval
to -1. But later when fast polling is disabled, saved poll
interval is restored (which will be default poll interval)
not -1. So polling is still ON.

So add a check to restore the poll interval only when
poll_timeout_msec != -1. This way polling will be still disabled.