Add power floor event processing
When the power floor event is received then 50% package wide idle
is injected as the power control is not enough to control thermal.

If the PCI bandwidth control cooling device is present, that is also
activated to control temperature.

This feature is activate only --power-floor-enable is passed from
command line.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>